+++ nms/fallback/src/main/java/com/zenya/damageindicator/nms/fallback/ProtocolNMSImpl.java
@@ -9,8 +9,8 @@ import org.bukkit.entity.LivingEntity;
 public class ProtocolNMSImpl implements ProtocolNMS {
 
     @Override
-    public Hologram getHologram(LivingEntity ent, String text) {
-        return new HologramImpl(ent, text);
+    public Hologram getHologram(LivingEntity ent, Location loc, String text) {
+        return new HologramImpl(ent, loc, text);
     }
 
     public static class HologramImpl implements Hologram {
@@ -19,16 +19,16 @@ public class ProtocolNMSImpl implements ProtocolNMS {
         private final LivingEntity entity;
         private final String text;
 
-        public HologramImpl(LivingEntity entity, String text) {
+        public HologramImpl(LivingEntity entity, Location loc, String text) {
             this.entity = entity;
             this.text = text;
         }
 
         @Override
-        public Hologram spawn(double offset, double speed, long duration) {
+        public Hologram spawn(double offsetX, double offsetY, double offsetZ, double speed, long duration) {
             sendCreatePacket();
             sendMetaPacket();
-            new HologramRunnable(this, entity, offset, speed, duration).start();
+            new HologramRunnable(this, entity, offsetX, offsetY, offsetZ, speed, duration).start();
             return this;
         }